Deleted 'vector.py' and renamed geodataset to vector
